Stub optimization routine classes for DAKOTA interface
