--- interfaces/optimization_dakota.py
@@ -1,20 +1,0 @@
-import numpy as np
-
-from interfaces.shape import Shape
-from interfaces.position import Position
-from interfaces.equation_of_motion import EquationOfMotion
-
-
-class OptimizationRoutineDakota(OptimizationRoutine):
-    """
-    An optimization interface for Dakota.
-    """
-    def __init__(self,parameters):
-        # self.parameter0 = parameters[0]
-        # ecc.
-        pass
-
-    def optimize(self):
-        # solution = ....
-        # return solution
-        pass
